+++ applications/stable-diffusion/instruct-pix2pix/01_trainer_pix2pix.sh
@@ -6,7 +6,6 @@ export DATASET_ID="fusing/instructpix2pix-1000-samples"
 accelerate launch --mixed_precision="fp16" train_instruct_pix2pix.py \
     --pretrained_model_name_or_path=$MODEL_NAME \
     --dataset_name=$DATASET_ID \
-    --enable_xformers_memory_efficient_attention \
     --resolution=256 --random_flip \
     --train_batch_size=4 --gradient_accumulation_steps=4 --gradient_checkpointing \
     --max_train_steps=15000 \
